legacy/backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmaes==0.10.0
colorlog==6.7.0
Cython==0.29.36
dataclasses-json==0.5.14
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
greenlet==2.0.2
hdbscan==0.8.33
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.285
langsmith==0.0.35
llvmlite==0.40.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
openai==0.28.0
optuna==3.3.0
packaging==23.1
Pillow==10.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pynndescent==0.5.10
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
SQLAlchemy==2.0.20
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
typing-inspect==0.9.0
typing_extensions==4.7.1
umap-learn==0.5.3
urllib3==2.0.4
yarl==1.9.2
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
bertopic==0.15.0
black==23.9.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmaes==0.10.0
colorlog==6.7.0
Cython==0.29.36
dataclasses-json==0.5.14
filelock==3.12.3
Flask==2.3.3
frozenlist==1.4.0
fsspec==2023.9.0
greenlet==2.0.2
hdbscan==0.8.33
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.285
langsmith==0.0.35
llvmlite==0.40.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
openai==0.28.0
optuna==3.3.0
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.16.1
pydantic==2.3.0
pydantic_core==2.6.3
pynndescent==0.5.10
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
SQLAlchemy==2.0.20
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.3
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
